move KFFittingSmoother.h in to its source file
